backend/requirements.txt
aiokafka==0.12.0
alembic==1.13.1
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
async-timeout==5.0.1
asyncpg==0.29.0
bcrypt==4.0.1
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.5
click==8.1.8
cloudpickle==3.1.2
contourpy==1.3.0
coverage==7.10.7
cryptography==46.0.5
cycler==0.12.1
dnspython==2.7.0
ecdsa==0.19.1
email-validator==2.3.0
exceptiongroup==1.3.1
fastapi==0.110.0
fonttools==4.60.2
greenlet==3.2.5
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.26.0
idna==3.11
imbalanced-learn==0.12.4
importlib_resources==6.5.2
iniconfig==2.1.0
joblib==1.5.3
kiwisolver==1.4.7
lightgbm==4.6.0
llvmlite==0.43.0
Mako==1.3.10
MarkupSafe==3.0.3
matplotlib==3.9.4
networkx==3.2.1
numba==0.60.0
numpy==1.26.4
packaging==26.0
pandas==2.2.0
passlib==1.7.4
pillow==11.3.0
pluggy==1.6.0
prometheus-client==0.19.0
prometheus-fastapi-instrumentator==6.1.0
psycopg2-binary==2.9.11
pyasn1==0.6.2
pycparser==2.23
pydantic==2.12.5
pydantic-settings==2.11.0
pydantic_core==2.41.5
pyparsing==3.3.2
pytest==8.0.0
pytest-asyncio==0.23.8
pytest-cov==4.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-jose==3.3.0
python-multipart==0.0.9
pytz==2026.1.post1
PyYAML==6.0.3
redis==5.0.1
requests==2.32.5
rsa==4.9.1
scikit-learn==1.4.0
scipy==1.13.1
seaborn==0.13.2
shap==0.44.1
six==1.17.0
slicer==0.0.7
sniffio==1.3.1
SQLAlchemy==2.0.27
starlette==0.36.3
threadpoolctl==3.6.0
tomli==2.4.0
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
uvicorn==0.27.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==12.0
xgboost==2.0.3
zipp==3.23.0
pytest==8.0.0
pytest-asyncio==0.23.8
httpx==0.27.0
